se  cambian tipos de datos, de int a decimal los atributos que tienen que ver con dinero, tambien se agrega script con BD para el uso hasta la sesion de reportes (80% )

+++ ManejoPresupuesto/ManejoPresupuesto/Models/Cuenta.cs
@@ -12,7 +12,7 @@ namespace ManejoPresupuesto.Models
         public  string Nombre { get; set; }
         [Display(Name = "Tipo Cuenta")]
         public int TipoCuentaId { get; set; }
-        public int Balance { get; set; }
+        public decimal Balance { get; set; }
         [StringLength(maximumLength: 1000)]
         public string Descripcion { get; set; }
         public string TipoCuenta { get; set; }
