+++ ManejoPresupuesto/ManejoPresupuesto/Models/IndiceCuentasViewModel.cs
@@ -4,6 +4,6 @@
     {
         public string TipoCuenta { get; set; }
         public IEnumerable<Cuenta> Cuentas { get; set; }
-        public int Balance => Cuentas.Sum(x => x.Balance);
+        public decimal Balance => Cuentas.Sum(x => x.Balance);
     }
 }
