+++ ManejoPresupuesto/ManejoPresupuesto/Models/ReporteMensualViewModel.cs
@@ -3,9 +3,9 @@
     public class ReporteMensualViewModel
     {
         public IEnumerable<ResultadoObtenerPorMes> TransaccionesPorMes { get; set; }
-        public int Ingresos => TransaccionesPorMes.Sum(x => x.Ingreso);
-        public int Gastos => TransaccionesPorMes.Sum(x => x.Gasto);
-        public int Total => Ingresos - Gastos;
+        public decimal Ingresos => TransaccionesPorMes.Sum(x => x.Ingreso);
+        public decimal Gastos => TransaccionesPorMes.Sum(x => x.Gasto);
+        public decimal Total => Ingresos - Gastos;
         public int Año { get; set; }
     }
 }
