+++ ManejoPresupuesto/ManejoPresupuesto/Models/ReporteSemanalViewModel.cs
@@ -2,9 +2,9 @@
 {
     public class ReporteSemanalViewModel
     {
-        public int Ingresos => TransaccionesPorSemana.Sum(x => x.Ingresos);
-        public int  Gastos => TransaccionesPorSemana.Sum(x => x.Gastos);
-        public int Total { get; set; }
+        public decimal Ingresos => TransaccionesPorSemana.Sum(x => x.Ingresos);
+        public decimal  Gastos => TransaccionesPorSemana.Sum(x => x.Gastos);
+        public decimal Total { get; set; }
         public DateTime FechaReferencia { get; set; }
         public IEnumerable<ResultadoObtenerPorSemana> TransaccionesPorSemana { get; set; }
     }
