+++ ManejoPresupuesto/ManejoPresupuesto/Models/ResultadoObtenerPorMes.cs
@@ -4,9 +4,9 @@
     {
         public int Mes { get; set; }
         public DateTime FechaReferencia { get; set; }
-        public int Monto { get; set; }
-        public int Ingreso { get; set; }
-        public int Gasto { get; set; }
+        public decimal Monto { get; set; }
+        public decimal Ingreso { get; set; }
+        public decimal Gasto { get; set; }
         public TipoOperacion TipoOperacionId { get; set; }
     }
 }
